feat(api): blocking client
