+++ api/src/client/mod.rs
@@ -1,12 +1,16 @@
-use reqwest::{IntoUrl, Url};
-use serde::{de::DeserializeOwned, Deserialize, Serialize};
+use serde::{Deserialize, Serialize};
 
-use crate::{
-    client::Result,
-    rpc::{ApiError, ApiResult, Request, ResponseObject},
-};
+use crate::rpc::{ApiError, ApiResult};
 
 mod_use::mod_use![error];
+
+#[cfg(feature = "client")]
+mod non_blocking;
+#[cfg(feature = "client")]
+pub use non_blocking::*;
+
+#[cfg(feature = "client_blocking")]
+pub mod blocking;
 
 #[derive(Serialize, Deserialize)]
 #[serde(untagged)]
@@ -21,37 +25,5 @@ impl<T> From<Shim<T>> for ApiResult<T> {
             Shim::Ok(res) => ApiResult::Ok(res),
             Shim::Err(err) => ApiResult::Err(err),
         }
-    }
-}
-
-pub struct Client {
-    client: reqwest::Client,
-    url: Url,
-}
-
-impl Client {
-    pub fn new(url: impl IntoUrl) -> Result<Self> {
-        Ok(Self {
-            client: reqwest::Client::new(),
-            url: url.into_url()?,
-        })
-    }
-
-    pub async fn invoke<R>(&self, req: R) -> Result<ApiResult<R::Res>>
-    where
-        R: Request + Serialize,
-        R::Res: DeserializeOwned,
-    {
-        Ok(self
-            .client
-            .post(self.url.clone())
-            .body(req.packed().to_json())
-            .header("Content-Type", "application/json")
-            .send()
-            .await?
-            .json::<ResponseObject<Shim<R::Res>>>()
-            .await?
-            .data
-            .into())
     }
 }
